% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pictures.R
\name{download_pictures}
\alias{download_pictures}
\title{Download the google image pictures for all titles of a given month of a year.}
\usage{
download_pictures(dir, month, year, n_movies = 5, movies_only = TRUE,
  n_pictures = 5, subdir_organize = FALSE)
}
\arguments{
\item{dir}{Character string pointing to a directory to save all the results.
Within this directory a new folder will be created called
moviepics-month-year.}

\item{month}{Integer or character month.}

\item{year}{Integer year.}

\item{n_movies}{Integer indicating the number of movies per week to get
images for.}

\item{movies_only}{Boolean indicator if only movie titles should be returned.
Exclude TV series like titles.}

\item{n_pictures}{Integer number indicating how many images to download from
the search. Capped at 20.}

\item{subdir_organize}{Boolean indicator if the movie image file should be
organized into further subdirectories by release date.}
}
\description{
Uses \link{get_titles} to get the dates and titles for the given month of the
year. Then uses \link{download_pictures_title} to download the images for
each title.
}
\examples{
download_pictures(".", 1, 2018)
download_pictures("..", "january" , 2018, subdir_organize = TRUE)

}
