% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pictures_title.R
\name{download_pictures_title}
\alias{download_pictures_title}
\title{Download the google image pictures for a title}
\usage{
download_pictures_title(dir, title, n_pictures = 5)
}
\arguments{
\item{dir}{Character string path to save the images.}

\item{title}{Character string of the movie title. The google image search is
actually for \code{paste(title, "movie poster")}.}

\item{n_pictures}{Integer number indicating how many images to download from
the search. Capped at 20.}
}
\description{
Saves image files for a single movie title. Uses
\href{http://phantomjs.org/}{phantomjs} to load the google image results and
downloads the selected amount of jpeg images for that title.
}
\examples{
download_pictures_title(".", "The Shape of Water")
download_pictures_title(".", "The Shape of Water", n_pictures = 10)

}
