% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_titles.R
\name{get_titles}
\alias{get_titles}
\title{Get movies released on DVD/Blu-ray for specified month and year.}
\usage{
get_titles(month, year, n_movies = 5, movies_only = TRUE)
}
\arguments{
\item{month}{Integer or character month.}

\item{year}{Integer year.}

\item{n_movies}{Integer number of movies to include for each week of releases in the month.}

\item{movies_only}{Boolean indicator if only movie titles should be returned.
Exclude TV series like titles.}
}
\value{
List of character vectors of the movies released. List is structured
  by release date.
}
\description{
Get the movies released for the month and year specified. Can optionally
include titles that look like TV series as well. Scrapes
\href{https://www.dvdsreleasedates.com/}{DVD Release Dates} for titles and
dates.
}
\examples{
get_titles("JAN", 2018)
get_titles(1, 2018)

}
